Fix breaks in responsive layouts
Using min-width breaks responsive layouts.

Here's an example of what the default Twenty twelve theme looks like before the fix on a mobile:
https://dl.dropboxusercontent.com/u/2758854/2013-09-30%2019.57.33.png

The only downside of this fix is that the help button gets a little squished on mobile.

Hiding the undo/redo buttons on mobile with a media query would fix it permanently. Do you want me to send a PR for this?